PrismPioneer_Backend/requirements.txt
Django>=5.0.0
djangorestframework>=3.14.0
drf-spectacular>=0.27.0
django-cors-headers>=4.3.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
django-environ>=0.11.0
Pillow>=10.0.0
markdown>=3.5.0
supabase>=2.72.9
